Move RewriteDataFilesProcedure to procedure package
And fix the error message thrown when target table does not exist
